% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{au_scoring}
\alias{au_scoring}
\title{The default scores for each outcome}
\usage{
au_scoring()
}
\value{
A named list with components \code{ace}, \code{serve_error}, \code{pass}, \code{pass_error}, \code{assist}, \code{set_error}, \code{kill}, \code{attack_error}, \code{block}, and \code{dig}, giving the numeric score for each
}
\description{
The default scores for each outcome
}
\references{
https://auprosports.com/volleyball/how-we-play-volleyball/
}
